refactor: add exception package
