+++ src/main/java/edu/fiuba/algo3/modelo/Defense/OperationalState.java
@@ -1,6 +1,7 @@
 package edu.fiuba.algo3.modelo.Defense;
 
 import edu.fiuba.algo3.modelo.Enemy.Enemy;
+import edu.fiuba.algo3.modelo.Exceptions.EnemyIsOutOfRange;
 
 public class OperationalState implements State{
     @Override
