+++ src/main/java/edu/fiuba/algo3/modelo/Exceptions/EnemyIsOutOfRange.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo.Defense;
+package edu.fiuba.algo3.modelo.Exceptions;
 
 public class EnemyIsOutOfRange extends RuntimeException {
 }
