+++ src/main/java/edu/fiuba/algo3/modelo/Exceptions/InsufficientCredits.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo;
+package edu.fiuba.algo3.modelo.Exceptions;
 
 public class InsufficientCredits extends RuntimeException{
 }
