+++ src/main/java/edu/fiuba/algo3/modelo/Exceptions/InvalidPlayersName.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo.Player;
+package edu.fiuba.algo3.modelo.Exceptions;
 
 public class InvalidPlayersName extends RuntimeException{
 }
