+++ src/main/java/edu/fiuba/algo3/modelo/Exceptions/PlayerIsDeadGameOver.java
@@ -1,3 +1,3 @@
-package edu.fiuba.algo3.modelo.Player;
+package edu.fiuba.algo3.modelo.Exceptions;
 
 public class PlayerIsDeadGameOver extends RuntimeException{ }
