+++ src/main/java/edu/fiuba/algo3/modelo/Exceptions/TowerIsUnderConstruction.java
@@ -1,3 +1,3 @@
-package edu.fiuba.algo3.modelo.Defense;
+package edu.fiuba.algo3.modelo.Exceptions;
 
 public class TowerIsUnderConstruction extends RuntimeException{ }
