+++ src/main/java/edu/fiuba/algo3/modelo/GameBoard/NotOccupiable.java
@@ -1,6 +1,6 @@
 package edu.fiuba.algo3.modelo.GameBoard;
 
-import edu.fiuba.algo3.modelo.Defense.TowerIsUnderConstruction;
+import edu.fiuba.algo3.modelo.Exceptions.NonConstructibleArea;
 
 public class NotOccupiable implements Occupiable{
     @Override
