+++ src/main/java/edu/fiuba/algo3/modelo/GameBoard/Stone.java
@@ -1,8 +1,5 @@
 package edu.fiuba.algo3.modelo.GameBoard;
 
-import edu.fiuba.algo3.modelo.Defense.EnemyIsOutOfRange;
-import edu.fiuba.algo3.modelo.Defense.TowerIsUnderConstruction;
-
 public class Stone extends Plot{
     public Stone(){state = new NotOccupiable();}
